requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
python-jose
python-dotenv
httpx
sse-starlette
sqlalchemy
aiosqlite
asyncpg
alembic
redis
celery
flower
apscheduler
pandas
numpy
scikit-learn
xgboost
lightgbm
catboost
optuna
shap
plotly
kaleido
seaborn
matplotlib
jinja2
pdfkit
prometheus-client
psutil
aiofiles
python-json-logger
pytest
pytest-asyncio
pytest-cov
black
isort
mypy
flake8
sphinx
passlib
bcrypt
python-jose[cryptography]
python-multipart
tqdm
joblib
PyYAML
python-dateutil
pytz
aiosmtplib
uvloop
orjson
email-validator
python-magic